add CurrentGeneration to tell which rsm.generation the underlying sts corresponds to
