--- internal/controller/rsm/stateful_set_revision_getter.go
@@ -1,193 +1,0 @@
-/*
-Copyright (C) 2022-2023 ApeCloud Co., Ltd
-
-This file is part of KubeBlocks project
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-package rsm
-
-import (
-	"context"
-	"encoding/json"
-
-	appsv1 "k8s.io/api/apps/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	roclient "github.com/apecloud/kubeblocks/internal/controller/client"
-)
-
-// controllerKind contains the schema.GroupVersionKind for this controller type.
-var controllerKind = appsv1.SchemeGroupVersion.WithKind("StatefulSet")
-
-var patchCodec = scheme.Codecs.LegacyCodec(appsv1.SchemeGroupVersion)
-
-type StatefulSetRevisionGetter interface {
-	// GetStatefulSetRevisions returns the current and update ControllerRevisions for set. It also
-	// returns a collision count that records the number of name collisions set saw when creating
-	// new ControllerRevisions. This count is incremented on every name collision and is used in
-	// building the ControllerRevision names for name collision avoidance. This method expects that
-	// revisions is sorted when supplied.
-	GetStatefulSetRevisions(set *appsv1.StatefulSet) (*appsv1.ControllerRevision, *appsv1.ControllerRevision, int32, error)
-}
-
-type realStatefulSetRevisionGetter struct {
-	ControllerHistoryInterface
-}
-
-var _ StatefulSetRevisionGetter = &realStatefulSetRevisionGetter{}
-
-type getterClient struct {
-	roclient.ReadonlyClient
-	client.Writer
-	client.StatusClient
-	client.SubResourceClientConstructor
-}
-
-func (g *getterClient) Scheme() *runtime.Scheme {
-	panic("should not be called")
-}
-
-func (g *getterClient) RESTMapper() meta.RESTMapper {
-	panic("should not be called")
-}
-
-var _ client.Client = &getterClient{}
-
-func NewStatefulSetRevisionGetter(ctx context.Context, readonlyClient roclient.ReadonlyClient) StatefulSetRevisionGetter {
-	cli := &getterClient{
-		ReadonlyClient: readonlyClient,
-		Writer: nil,
-		StatusClient: nil,
-		SubResourceClientConstructor: nil,
-	}
-	return &realStatefulSetRevisionGetter{
-		ControllerHistoryInterface: NewHistory(ctx, cli),
-	}
-}
-
-// getPatch returns a strategic merge patch that can be applied to restore a StatefulSet to a
-// previous version. If the returned error is nil the patch is valid. The current state that we save is just the
-// PodSpecTemplate. We can modify this later to encompass more state (or less) and remain compatible with previously
-// recorded patches.
-func getPatch(set *appsv1.StatefulSet) ([]byte, error) {
-	data, err := runtime.Encode(patchCodec, set)
-	if err != nil {
-		return nil, err
-	}
-	var raw map[string]interface{}
-	err = json.Unmarshal(data, &raw)
-	if err != nil {
-		return nil, err
-	}
-	objCopy := make(map[string]interface{})
-	specCopy := make(map[string]interface{})
-	spec := raw["spec"].(map[string]interface{})
-	template := spec["template"].(map[string]interface{})
-	specCopy["template"] = template
-	template["$patch"] = "replace"
-	objCopy["spec"] = specCopy
-	patch, err := json.Marshal(objCopy)
-	return patch, err
-}
-
-// nextRevision finds the next valid revision number based on revisions. If the length of revisions
-// is 0 this is 1. Otherwise, it is 1 greater than the largest revision's Revision. This method
-// assumes that revisions has been sorted by Revision.
-func nextRevision(revisions []*appsv1.ControllerRevision) int64 {
-	count := len(revisions)
-	if count <= 0 {
-		return 1
-	}
-	return revisions[count-1].Revision + 1
-}
-
-// newRevision creates a new ControllerRevision containing a patch that reapplies the target state of set.
-// The Revision of the returned ControllerRevision is set to revision. If the returned error is nil, the returned
-// ControllerRevision is valid. StatefulSet revisions are stored as patches that re-apply the current state of set
-// to a new StatefulSet using a strategic merge patch to replace the saved state of the new StatefulSet.
-func newRevision(set *appsv1.StatefulSet, revision int64, collisionCount *int32) (*appsv1.ControllerRevision, error) {
-	patch, err := getPatch(set)
-	if err != nil {
-		return nil, err
-	}
-	cr, err := NewControllerRevision(set,
-		controllerKind,
-		set.Spec.Template.Labels,
-		runtime.RawExtension{Raw: patch},
-		revision,
-		collisionCount)
-	if err != nil {
-		return nil, err
-	}
-	if cr.ObjectMeta.Annotations == nil {
-		cr.ObjectMeta.Annotations = make(map[string]string)
-	}
-	for key, value := range set.Annotations {
-		cr.ObjectMeta.Annotations[key] = value
-	}
-	return cr, nil
-}
-
-func (g *realStatefulSetRevisionGetter) ListRevisions(set *appsv1.StatefulSet) ([]*appsv1.ControllerRevision, error) {
-	selector, err := metav1.LabelSelectorAsSelector(set.Spec.Selector)
-	if err != nil {
-		return nil, err
-	}
-	return g.ListControllerRevisions(set, selector)
-}
-
-// GetStatefulSetRevisions returns the current and update ControllerRevisions for set. It also
-// returns a collision count that records the number of name collisions set saw when creating
-// new ControllerRevisions. This count is incremented on every name collision and is used in
-// building the ControllerRevision names for name collision avoidance. This method expects that
-// revisions is sorted when supplied.
-func (g *realStatefulSetRevisionGetter) GetStatefulSetRevisions(set *appsv1.StatefulSet) (*appsv1.ControllerRevision, *appsv1.ControllerRevision, int32, error) {
-	var currentRevision, updateRevision *appsv1.ControllerRevision
-	revisions, err := g.ListRevisions(set)
-	SortControllerRevisions(revisions)
-
-	// Use a local copy of set.Status.CollisionCount to avoid modifying set.Status directly.
-	// This copy is returned so the value gets carried over to set.Status in updateStatefulSet.
-	var collisionCount int32
-	if set.Status.CollisionCount != nil {
-		collisionCount = *set.Status.CollisionCount
-	}
-
-	// create a new revision from the current set
-	updateRevision, err = newRevision(set, nextRevision(revisions), &collisionCount)
-	if err != nil {
-		return nil, nil, collisionCount, err
-	}
-
-	// attempt to find the revision that corresponds to the current revision
-	for i := range revisions {
-		if revisions[i].Name == set.Status.CurrentRevision {
-			currentRevision = revisions[i]
-			break
-		}
-	}
-
-	// if the current revision is nil we initialize the history by setting it to the update revision
-	if currentRevision == nil {
-		currentRevision = updateRevision
-	}
-
-	return currentRevision, updateRevision, collisionCount, nil
-}
