+++ internal/controller/rsm/types.go
@@ -50,6 +50,7 @@ const (
 
 	roleLabelKey          = "kubeblocks.io/role"
 	rsmAccessModeLabelKey = "rsm.workloads.kubeblocks.io/access-mode"
+	rsmGenerationLabelKey = "rsm.workloads.kubeblocks.io/controller-generation"
 
 	defaultPodName = "Unknown"
